Define a class to store runtime object
The scanner can convert the textual representation of a value
into a runtime object that can be easily used by the interpreter
to evaluate the value. However, Cpp does not have a sophisticated
type to store the object. Hence, I'm defining my own class to
store both the value and textual representation.