Add rule to compile and run test file
Now I can compile and run test files using the same make command
I use to build the interpreter. I have started to call the make
command from Emacs and it has made my workflow much better as I
don't have to switch between the terminal and editor anymore.

In the process of writing the front-end for the interpreter, I
have christened the interpreter as Dlox because I'm writing this
interpreter in C++ and technically C++ == D and lox comes from the
name of original interpreter written by Bob Nystorm.