Handle parse errors gracefully
Previously, whenever the parser encountered a line without
semicolon at the end, it would throw an error and crash the
entire program as there was no mechanism to handle the errors.
Now I have wrapped the call to parse statements in a try and
catch block and it can handle errors by only printing the
appropriate error message.