Add blocks to Dlox
Now variables are lexically scoped using blocks ({}). This
implementation is identical to how I implemented in Scheme
evaluator: Every instance of an environment has a pointer to the
enclosing environment. If the variable being searched for is not
found in the current environment then the program searches for it
in the previous environment recursively until it finds a reference
and this recursive search hits the base at at the global
environment.

Similarly, for executing statements inside a block, the
current environment becomes the enclosing environment and a
new (empty) environment becomes the current environment. After
executing all the statements inside the block, the current
environment is deleted and the enclosing environment becomes the
current environment.