Change the type of number from int to double
Previously, I casted every number to Cpp's int. So there was
no way to represent float in Dlox and it resulted in incorrect
values when evaluating binary and unary expressions.

To fix this, I'm casting every number to Cpp's double. This allows
me to represent both int and float by using a small trick to trim
off the leading zeros.