Bad implementation to check for unused variables
I thought define() was used when a variable's value was looked
up but it turns out define() is used to mark a variable as
ready for usage. Declaring and defining are split into two parts
to prevent declaring and initializing local variable with its own
name. For example, var a = a in a local scope where a is declared
already either in global or local scope. So using declare() and
define() to figure out if variable is used or not is not a good
idea.

A better way to implement this would be to check the data
resolver gives to the interpreter. Resolver gives the distance
between every local variable expression and its value in memory
map. So if there is no distance for a local variable expression,
it signifies that the variable is never used locally and that
should be a compile warning/error.