Primitive approach to check for unused variables
Every time a variable is declared, its name is added to map
with its name as key and zero as value. I add one to its value
whenever the variable gets resolved and before exiting a scope,
I loop through all the key value pair in the map and throw error
if any value is zero(was never resolved).

While this approach works, it could be implemented in a better
way by storing the Token in the map instead of the token's name.
I'm also concerned about the added cost as it has to iterate
through every variable declared before exiting a scope.