Move visitor interface to a new file
Lambda expressions have statements in their body but it was
impossible to declare statements a expression as the header file
of statements include expressions and there is no way for
expressions to include statements. So I moved both the visitor
interface to a new file which will be included by both statements
and expressions. Now both statements and expression have access
to each other.