Lambda expression works
It took me so long to figure out the bug because there was no
bug and everything in Interpreter was implemented correctly.
However, to make Lambdas work quickly, I had completely ignored
Resolver and forgot that Interpreter heavily depends on it to
know which variable to fetch. So since resolver was not giving
any data about the variables inside a lamda expression's scope,
the interpreter assumed there was no local variables in the scope.
Hence, it went to global environment to search for the
parameter's values.