Make Lists can be subscriptable
I'm thrilled to write that Dlox can now primitively output an
element from the list based on its index. This is a remarkable
milestone for Dlox.

However, there are a couple of things that needs to be fixed.
Firstly, Dlox parses every number as double but for indexing a
list, we stricly need an int. So I should figure out a way to
gracefully cast double to int. Lastly, there's no checks for
indexing out of bounds. So this results in g++ throwing a nasty
message and Dlox terminates.