Use Subscript expression for both storing and getting list elements
Previously, I was using two different expression nodes to handle
set and get list elements. But later on I realized there was only
a slight difference between both the expressions(value). So now
if the subscript expression has a value, then the value is set at
the given index in the list but if it does not have a value, then
the value at the given index is returned.

Though the idea sounds good, my implementation looks ugly with a
lot of if else conditions.