Make resolver erase used variables from identifiers hashmap
Previously, for checking unused variables,
the identifiers variable contained a vector of hashmaps. Each
hashmap would contain the variables declared in a scope as
pairs of variable name and number of times it was used in the
scope. Finally, the resolver would iterate through all the
elements in a hashmap and throw a warning on pairs whose
value is zero as this suggests that variable was never used in
the scope.

Now, to improve on this idea, whenever a variable is evaluated
in a scope, I erase it from the hashmap. This makes it easy for
the resolver as it can throw an error on all values in a hashmap
without checking if its value is zero.